feat(forms/shaping): move composeExplanation into service
This function iterates commands and humanizes them with progressive
state tracking — it is shaping business logic, not route logic.
