+++ src/services/forms/index.ts
@@ -31,7 +31,7 @@ export { createReviewService } from './review'
 export type { Command, ProjectState } from './shaping/commands'
 export { commandSchema } from './shaping/commands'
 export { executeBatch } from './shaping/executor'
-export { humanize } from './shaping/humanize'
+export { composeExplanation, humanize } from './shaping/humanize'
 export { createShapingRegistry } from './shaping/registry'
 export { commandTools } from './shaping/tools'
 export type { FormShaper } from './shaping/types'
